pyproject.toml
[project]
name = "undriven"
version = "0.1.0"
description = "Track car miles replaced by non-car trips"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.0",
    "gspread>=6.0",
    "google-auth>=2.0",
    "requests>=2.31",
]

